schemachangerccl,sctest,scdecomp: enrich schemachanger CCL tests
This commit adds more data-driven tests to cover CCL features support by
the declarative schema changer.

Release note: None
