+++ pkg/sql/schemachanger/scdecomp/BUILD.bazel
@@ -36,28 +36,17 @@ go_test(
     ],
     data = glob(["testdata/**"]),
     deps = [
-        ":scdecomp",
         "//pkg/base",
-        "//pkg/kv",
         "//pkg/security",
         "//pkg/security/securitytest",
         "//pkg/server",
-        "//pkg/sql",
-        "//pkg/sql/catalog",
-        "//pkg/sql/catalog/descs",
-        "//pkg/sql/parser",
-        "//pkg/sql/schemachanger/scdeps/sctestutils",
-        "//pkg/sql/schemachanger/scpb",
-        "//pkg/sql/sem/catid",
-        "//pkg/sql/sem/tree",
+        "//pkg/sql/schemachanger/scrun",
+        "//pkg/sql/schemachanger/sctest",
         "//pkg/testutils",
         "//pkg/testutils/serverutils",
-        "//pkg/testutils/sqlutils",
         "//pkg/testutils/testcluster",
         "//pkg/util/leaktest",
         "//pkg/util/log",
         "//pkg/util/randutil",
-        "@com_github_cockroachdb_datadriven//:datadriven",
-        "@com_github_stretchr_testify//require",
     ],
 )
